final for wednesday
